Move the shutdown handler into io package.
